+++ cinema/Color.h
@@ -3,16 +3,16 @@ class Color
 {
 public:
 	Color();
-	Color(float rr, float gg, float bb);
-	Color(float intensity);
+	Color(double rr, double gg, double bb);
+	Color(double intensity);
 	~Color();
 	Color operator+(const Color &c)const;
 	Color & operator+=(const Color &c);
 	Color operator*(const Color &c)const;
 	Color & operator*=(const Color &c);
-	Color operator*(const float f)const;
+	Color operator*(const double f)const;
 public:
-	float r, g, b;
+	double r, g, b;
 	static const Color BLACK;
 	static const Color WHITE;
 };
