+++ cinema/Geometry.h
@@ -6,6 +6,6 @@ class Geometry
 public:
 	Geometry();
 	virtual ~Geometry() = 0;
-	virtual bool Intersect(const Ray &ray, float *t, Intersection *intersection) = 0;
+	virtual bool Intersect(const Ray &ray, double *t, Intersection *intersection) = 0;
 };
