+++ cinema/Point.h
@@ -4,12 +4,12 @@ class Point
 {
 public:
 	Point();
-	Point(float xx, float yy, float zz);
+	Point(double xx, double yy, double zz);
 	~Point();
 	Point operator+(const Vector &v) const;
 	Point operator-(const Vector &v) const;
 	Vector operator+(const Point &p) const;
 	Vector operator-(const Point &p) const;
-	float x, y, z;
+	double x, y, z;
 };
