+++ cinema/Triangle.cpp
@@ -12,25 +12,25 @@ Triangle::~Triangle()
 {
 }
 
-bool Triangle::Intersect(const Ray &ray, float *t, Intersection *intersection)
+bool Triangle::Intersect(const Ray &ray, double *t, Intersection *intersection)
 {
 	Vector s = ray.o - p0;
 	Vector e1 = p1 - p0;
 	Vector e2 = p2 - p0;
 	Vector s1 = Vector::Cross(ray.d, e2);
 	Vector s2 = Vector::Cross(s, e1);
-	float determinantOfCoefficient = Vector::Dot(s1, e1);
+	double determinantOfCoefficient = Vector::Dot(s1, e1);
 	//TODO 0.0f may be a problem
 	if (determinantOfCoefficient == 0.0f)
 	{
 		return false;
 	}
-	float b1 = Vector::Dot(s1, s) / determinantOfCoefficient;
+	double b1 = Vector::Dot(s1, s) / determinantOfCoefficient;
 	if (b1 > 1.0f || b1 < 0.0f)
 	{
 		return false;
 	}
-	float b2 = Vector::Dot(s2, ray.d) / determinantOfCoefficient;
+	double b2 = Vector::Dot(s2, ray.d) / determinantOfCoefficient;
 	if (b2 > 1.0f || b2 < 0.0f)
 	{
 		return false;
